requirements-security.txt
# Security-focused requirements for Telecom KPI Dashboard
# Install these in addition to the base requirements.txt

# Core security libraries
cryptography>=3.4.8
bcrypt>=3.2.0
python-dotenv>=0.19.0

# Input validation and sanitization
bleach>=4.1.0
html5lib>=1.1

# Rate limiting and security headers
slowapi>=0.1.5
secure>=0.3.0

# Logging and monitoring
structlog>=21.1.0
python-json-logger>=2.0.0

# Environment and configuration security
pydantic>=1.8.0
pyyaml>=6.0

# Additional security tools for production
bandit>=1.7.0  # Security linter
safety>=1.10.0  # Vulnerability scanner
